refactor(patina): split return_in_computed_property tests into submodule
